Cleaning up Compiler warnings

+++ GUITimer.cpp
@@ -52,7 +52,7 @@ void FrameRateCapper::cap_frame_rate(){
 FrameRateCapper::~FrameRateCapper(){
 	if (!capping) return;
 	
-	if (fps.get_time() < 1000 / fps_cap) {
+	if (fps.get_time() < (Uint32) (1000 / fps_cap)) {
 		SDL_Delay((1000 / fps_cap) - fps.get_time());
 	}
 }
